+++ sdk/network/azure-resourcemanager-network/src/main/java/com/azure/resourcemanager/network/implementation/NatGatewaysClientImpl.java
@@ -177,7 +177,7 @@ public final class NatGatewaysClientImpl implements InnerSupportsGet<NatGatewayI
         if (natGatewayName == null) {
             return Mono.error(new IllegalArgumentException("Parameter natGatewayName is required and cannot be null."));
         }
-        final String apiVersion = "2025-05-01";
+        final String apiVersion = "2025-07-01";
         final String accept = "application/json";
         return FluxUtil
             .withContext(context -> service.getByResourceGroup(this.client.getEndpoint(), apiVersion,
@@ -216,7 +216,7 @@ public final class NatGatewaysClientImpl implements InnerSupportsGet<NatGatewayI
         if (natGatewayName == null) {
             return Mono.error(new IllegalArgumentException("Parameter natGatewayName is required and cannot be null."));
         }
-        final String apiVersion = "2025-05-01";
+        final String apiVersion = "2025-07-01";
         final String accept = "application/json";
         context = this.client.mergeContext(context);
         return service.getByResourceGroup(this.client.getEndpoint(), apiVersion, this.client.getSubscriptionId(),
@@ -308,7 +308,7 @@ public final class NatGatewaysClientImpl implements InnerSupportsGet<NatGatewayI
         } else {
             parameters.validate();
         }
-        final String apiVersion = "2025-05-01";
+        final String apiVersion = "2025-07-01";
         final String contentType = "application/json";
         final String accept = "application/json";
         return FluxUtil
@@ -353,7 +353,7 @@ public final class NatGatewaysClientImpl implements InnerSupportsGet<NatGatewayI
         } else {
             parameters.validate();
         }
-        final String apiVersion = "2025-05-01";
+        final String apiVersion = "2025-07-01";
         final String contentType = "application/json";
         final String accept = "application/json";
         context = this.client.mergeContext(context);
@@ -543,7 +543,7 @@ public final class NatGatewaysClientImpl implements InnerSupportsGet<NatGatewayI
         } else {
             parameters.validate();
         }
-        final String apiVersion = "2025-05-01";
+        final String apiVersion = "2025-07-01";
         final String contentType = "application/json";
         final String accept = "application/json";
         return FluxUtil
@@ -588,7 +588,7 @@ public final class NatGatewaysClientImpl implements InnerSupportsGet<NatGatewayI
         } else {
             parameters.validate();
         }
-        final String apiVersion = "2025-05-01";
+        final String apiVersion = "2025-07-01";
         final String contentType = "application/json";
         final String accept = "application/json";
         context = this.client.mergeContext(context);
@@ -675,7 +675,7 @@ public final class NatGatewaysClientImpl implements InnerSupportsGet<NatGatewayI
         if (natGatewayName == null) {
             return Mono.error(new IllegalArgumentException("Parameter natGatewayName is required and cannot be null."));
         }
-        final String apiVersion = "2025-05-01";
+        final String apiVersion = "2025-07-01";
         return FluxUtil
             .withContext(context -> service.delete(this.client.getEndpoint(), apiVersion,
                 this.client.getSubscriptionId(), resourceGroupName, natGatewayName, context))
@@ -711,7 +711,7 @@ public final class NatGatewaysClientImpl implements InnerSupportsGet<NatGatewayI
         if (natGatewayName == null) {
             return Mono.error(new IllegalArgumentException("Parameter natGatewayName is required and cannot be null."));
         }
-        final String apiVersion = "2025-05-01";
+        final String apiVersion = "2025-07-01";
         context = this.client.mergeContext(context);
         return service.delete(this.client.getEndpoint(), apiVersion, this.client.getSubscriptionId(), resourceGroupName,
             natGatewayName, context);
@@ -872,7 +872,7 @@ public final class NatGatewaysClientImpl implements InnerSupportsGet<NatGatewayI
             return Mono
                 .error(new IllegalArgumentException("Parameter resourceGroupName is required and cannot be null."));
         }
-        final String apiVersion = "2025-05-01";
+        final String apiVersion = "2025-07-01";
         final String accept = "application/json";
         return FluxUtil
             .withContext(context -> service.listByResourceGroup(this.client.getEndpoint(), apiVersion,
@@ -908,7 +908,7 @@ public final class NatGatewaysClientImpl implements InnerSupportsGet<NatGatewayI
             return Mono
                 .error(new IllegalArgumentException("Parameter resourceGroupName is required and cannot be null."));
         }
-        final String apiVersion = "2025-05-01";
+        final String apiVersion = "2025-07-01";
         final String accept = "application/json";
         context = this.client.mergeContext(context);
         return service
@@ -996,7 +996,7 @@ public final class NatGatewaysClientImpl implements InnerSupportsGet<NatGatewayI
             return Mono.error(new IllegalArgumentException(
                 "Parameter this.client.getSubscriptionId() is required and cannot be null."));
         }
-        final String apiVersion = "2025-05-01";
+        final String apiVersion = "2025-07-01";
         final String accept = "application/json";
         return FluxUtil
             .withContext(context -> service.list(this.client.getEndpoint(), apiVersion, this.client.getSubscriptionId(),
@@ -1026,7 +1026,7 @@ public final class NatGatewaysClientImpl implements InnerSupportsGet<NatGatewayI
             return Mono.error(new IllegalArgumentException(
                 "Parameter this.client.getSubscriptionId() is required and cannot be null."));
         }
-        final String apiVersion = "2025-05-01";
+        final String apiVersion = "2025-07-01";
         final String accept = "application/json";
         context = this.client.mergeContext(context);
         return service.list(this.client.getEndpoint(), apiVersion, this.client.getSubscriptionId(), accept, context)