+++ sdk/spring/spring-cloud-azure-appconfiguration-config/src/test/java/com/azure/spring/cloud/appconfiguration/config/implementation/AppConfigurationApplicationSettingPropertySourceSnapshotTest.java
@@ -45,8 +45,6 @@ import com.azure.data.appconfiguration.models.ConfigurationSetting;
 import com.azure.data.appconfiguration.models.FeatureFlagConfigurationSetting;
 import com.azure.spring.cloud.appconfiguration.config.implementation.http.policy.TracingInfo;
 import com.azure.spring.cloud.appconfiguration.config.implementation.properties.AppConfigurationProperties;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 
 
 public class AppConfigurationApplicationSettingPropertySourceSnapshotTest {
@@ -79,8 +77,6 @@ public class AppConfigurationApplicationSettingPropertySourceSnapshotTest {
     private static final ConfigurationSetting ITEM_NULL =
         createItem(KEY_FILTER, TEST_KEY_3, TEST_VALUE_3, TEST_LABEL_3, null);
 
-    private static final ObjectMapper MAPPER = new ObjectMapper();
-
     private List<ConfigurationSetting> testItems = new ArrayList<>();
 
     private AppConfigurationSnapshotPropertySource propertySource;
@@ -109,7 +105,6 @@ public class AppConfigurationApplicationSettingPropertySourceSnapshotTest {
     @BeforeEach
     public void init() {
         session = Mockito.mockitoSession().initMocks(this).strictness(Strictness.STRICT_STUBS).startMocking();
-        MAPPER.setPropertyNamingStrategy(PropertyNamingStrategies.KEBAB_CASE);
 
         MockitoAnnotations.openMocks(this);